Switch many syscalls to return longs

+++ libutil/include/libutil.h
@@ -17,6 +17,6 @@ void xwrite(int fd, const void *buf, size_t n);
 
 uint64_t now_usec(void);
 
-int setaffinity(int c);
+long setaffinity(int c);
 
 END_DECLS
