+++ stdinc/signal.h
@@ -7,6 +7,6 @@ BEGIN_DECLS
 typedef void (*sighandler_t)(int);
 
 sighandler_t signal(int sig, sighandler_t func);
-int sigaction(int sig, struct sigaction* act, struct sigaction* oact);
+long sigaction(int sig, struct sigaction* act, struct sigaction* oact);
 
 END_DECLS
